requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
moviepy==1.0.3
duckduckgo-search==4.1.1
phidata==2.3.5
pillow==10.2.0
pytube==15.0.0
SpeechRecognition